requirements.txt
pyttsx3
newsapi
SpeechRecognition
tinyDB
winshell
imdbpy
geocoder
PyDictionary
newsapi-python
wolframalpha